% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importar.R
\name{read_rais}
\alias{read_rais}
\title{Ler tabelas extraidas do BI da RAIS}
\usage{
read_rais(caminho)
}
\arguments{
\item{caminho}{Caminho para arquivo}
}
\value{
um data.frame
}
\description{
Ler tabelas extraidas do BI da RAIS
}
