% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importar.R
\name{tidy_rais}
\alias{tidy_rais}
\title{Arrumar dados da RAIS}
\usage{
tidy_rais(df)
}
\arguments{
\item{df}{O data.frame gerado por \link{read_rais}}
}
\value{
Um data.frame "tidy"
}
\description{
Arrumar dados da RAIS
}
